+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Changelog
 =========
 
-## Unreleased
+## 1.0.3
 
-Initial release.
+- Fixed a PHP error that occurred when subscribing a user to a plan. ([#17](https://github.com/craftcms/commerce-stripe/issues/17))
+
+## 1.0.2
+
+- Added support for card tokens, not just payment sources. ([#9](https://github.com/craftcms/commerce-stripe/issues/9))
+- Allow creating payment source on behalf of users. ([#12](https://github.com/craftcms/commerce-stripe/issues/12))
+
+## 1.0.1
+
+- Fix a bug with a missing dependency.
+
+## 1.0.0
+
+- Initial release.